+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/AttributesJson.kt
@@ -6,30 +6,27 @@ import android.util.JsonWriter
 import com.bugsnag.android.performance.Attributes
 
 internal fun JsonWriter.value(attributes: Attributes): JsonWriter {
-    beginArray()
+    return array {
+        attributes.forEach { (key, value) ->
+            obj {
+                name("key").value(key)
 
-    attributes.forEach { (key, value) ->
-        beginObject() // attribute
-            .name("key").value(key)
-            .name("value")
-
-        toAttributeJson(value, this)
-
-        endObject() // attribute
+                name("value")
+                toAttributeJson(value, this)
+            }
+        }
     }
-
-    return endArray()
 }
 
 private fun toAttributeJson(value: Any, json: JsonWriter) {
-    json.beginObject()
-    when (value) {
-        is String -> json.name("stringValue").value(value)
-        is Float -> json.name("doubleValue").value(value.toDouble())
-        is Double -> json.name("doubleValue").value(value)
-        is Boolean -> json.name("boolValue").value(value)
-        // int64 is JSON encoded as a String
-        is Long, Int, Short, Byte -> json.name("intValue").value(value.toString())
+    json.obj {
+        when (value) {
+            is String -> json.name("stringValue").value(value)
+            is Float -> json.name("doubleValue").value(value.toDouble())
+            is Double -> json.name("doubleValue").value(value)
+            is Boolean -> json.name("boolValue").value(value)
+            // int64 is JSON encoded as a String
+            is Long, Int, Short, Byte -> json.name("intValue").value(value.toString())
+        }
     }
-    json.endObject()
 }